+++ NEWS.md
@@ -1,14 +1,15 @@
 # spfilteR 1.0.0.9000
 
 - new vignette name
-- include `MI.local()`
+- include `MI.local()` function to calculate local Moran's I
 - update reference to `MI.local()` in documentation files
+- include `vp()` function for variation partitioning
 
 ---
 
 # spfilteR 1.0.0
 
-- include `MI.decomp()`
+- include `MI.decomp()` to decompose Moran's I
 - rename `MI.resid()`
 - prepare for CRAN submission